+++ arangod/VocBase/SingleServerTraverser.h
@@ -97,7 +97,7 @@ class SingleServerTraverser final : public Traverser {
   ///        Returns true if the vertex passes filtering conditions
 
   bool getSingleVertex(arangodb::velocypack::Slice, arangodb::velocypack::Slice,
-                       size_t depth, arangodb::velocypack::Slice&) override;
+                       uint64_t depth, arangodb::velocypack::Slice&) override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Function to fetch the real data of a vertex into an AQLValue
